feat: Correcao telefone

+++ Application/Application.Authentication/ViewModels/RegisterViewModel.cs
@@ -18,7 +18,7 @@ public class RegisterViewModel
 public class RegistrarPessoaViewModel
 {
     public string Nome { get; set; }
-    public int Telefone { get; set; }
+    public long Telefone { get; set; }
     public EnderecoViewModel Endereco { get; set; }
     public DateTime DataDeNascimento { get; set; }
     public SexoViewModel Sexo { get; set; }
